Update PrestaSharpDeserializer.cs
i think this throw is more in equation with the error managed in Prestashop, i try with a return of webservice and works. 

The webservice can give you error like this

<prestashop xmlns:xlink="http://www.w3.org/1999/xlink">
<errors>
<error>
<message>
<![CDATA[
Internal error. To see this error please display the PHP errors.
]]>
</message>
</error>
<error>
<message>
<![CDATA[
Internal error. To see this error please display the PHP errors.
]]>
</message>
</error>
<error>
<message>
<![CDATA[
Internal error. To see this error please display the PHP errors.
]]>
</message>
</error>
<error>
<message>
<![CDATA[
Internal error. To see this error please display the PHP errors.
]]>
</message>
</error>
</errors>
</prestashop>